examples/04-intermediate/chatbot-web-app/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai
gunicorn==21.2.0
